+++ src/theme/components/card.ts
@@ -9,23 +9,23 @@ const imageSizes = {
 const getBodyPaddingStyles = ({ border, hasImage, imageIsAtEnd, isRow }) => {
   let bodyPadding = null;
   if (border) {
-    bodyPadding = "s";
+    bodyPadding = "inset.default";
     if (hasImage) {
-      bodyPadding = "0 var(--nypl-space-s) var(--nypl-space-s)";
+      bodyPadding = "0 var(--nypl-space-inset-default) var(--nypl-space-inset-default)";
     }
   }
   if (isRow && border) {
-    bodyPadding = "var(--nypl-space-s)";
+    bodyPadding = "var(--nypl-space-inset-default)";
   }
   if (isRow && border && hasImage) {
     bodyPadding = {
-      base: "0 var(--nypl-space-s) var(--nypl-space-s)",
-      md: "var(--nypl-space-s) var(--nypl-space-s) var(--nypl-space-s) 0",
+      base: "0 var(--nypl-space-inset-default) var(--nypl-space-inset-default)",
+      md: "var(--nypl-space-inset-default) var(--nypl-space-inset-default) var(--nypl-space-inset-default) 0",
     };
     if (imageIsAtEnd) {
       bodyPadding = {
-        base: "var(--nypl-space-s) var(--nypl-space-s) 0",
-        md: "var(--nypl-space-s) 0 var(--nypl-space-s) var(--nypl-space-s)",
+        base: "var(--nypl-space-inset-default) var(--nypl-space-inset-default) 0",
+        md: "var(--nypl-space-inset-default) 0 var(--nypl-space-inset-default) var(--nypl-space-inset-default)",
       };
     }
   }