+++ src/theme/components/notification.ts
@@ -21,7 +21,7 @@ const Notification = {
       container: {
         margin: "auto",
         maxWidth: "var(--nypl-breakpoint-xl)",
-        padding: "s",
+        padding: "inset.default",
         paddingRight: dismissible ? "l" : null,
         paddingLeft: isCentered && dismissible ? "l" : null,
         width: "100%",