+++ src/theme/foundations/spacing.ts
@@ -1,12 +1,13 @@
 /**
- * This is based on:
+ * Chakra spacing values are based on:
  *   root size of 16px or 1rem
  *   1 spacing unit is 4px or 0.25rem
  *
  * Chakra's mental model: If you need a spacing of 40px, divide it by 4.
  * That'll give you 10. Then use it in your component.
  *
- * The DS uses eight CSS variables/values found in "src/styles/03-space/_space.css":
+ * The DS uses nine CSS variables/values found in "src/styles/03-space/_space.css":
+ *   --nypl-space-xxxs = 2px or 0.125rem
  *   --nypl-space-xxs = 4px or 0.25rem
  *   --nypl-space-xs = 8px or 0.5rem
  *   --nypl-space-s = 16px or 1rem
@@ -32,45 +33,27 @@
  *        16           |         xxl       |   --nypl-space-xxl
  *        24           |         xxxl      |   --nypl-space-xxxl
  */
-export const spacing = {
+const chakraSpacingTokens = {
   px: "1px",
   0: "0",
-  // --nypl-space-xxxs = 2px
-  xxxs: "0.125rem",
   0.5: "0.125rem",
-  // --nypl-space-xxs = 4px
-  xxs: "0.25rem",
   1: "0.25rem",
   1.5: "0.375rem",
-  // --nypl-space-xs = 8px
-  xs: "0.5rem",
   2: "0.5rem",
   2.5: "0.625rem",
   3: "0.75rem",
   3.5: "0.875rem",
-  // --nypl-space-s = 16px
-  s: "1rem",
   4: "1rem",
   5: "1.25rem",
-  // --nypl-space-m = 24px
-  m: "1.5rem",
   6: "1.5rem",
   7: "1.75rem",
-  // --nypl-space-l = 32px
-  l: "2rem",
   8: "2rem",
   9: "2.25rem",
   10: "2.5rem",
-  // --nypl-space-xl = 48px
-  xl: "3rem",
   12: "3rem",
   14: "3.5rem",
-  // --nypl-space-xxl = 64px
-  xxl: "4rem",
   16: "4rem",
   20: "5rem",
-  // --nypl-space-xxxl = 96px
-  xxxl: "6rem",
   24: "6rem",
   28: "7rem",
   32: "8rem",
@@ -85,4 +68,68 @@ export const spacing = {
   72: "18rem",
   80: "20rem",
   96: "24rem",
+};
+const reservoirSpacingValues = {
+  xxxs: "0.125rem", // 2px
+  xxs: "0.25rem", // 4px
+  xs: "0.5rem", // 8px
+  s: "1rem", // 16px
+  m: "1.5rem", // 24px
+  l: "2rem", // 32px
+  xl: "3rem", // 48px
+  xxl: "4rem", // 64px
+  xxxl: "6rem", // 96px
+};
+const reservoirSpacingTokens = {
+  // Reservoir generic tokens
+  xxxs: reservoirSpacingValues["xxxs"],
+  xxs: reservoirSpacingValues["xxs"],
+  xs: reservoirSpacingValues["xs"],
+  s: reservoirSpacingValues["s"],
+  m: reservoirSpacingValues["m"],
+  l: reservoirSpacingValues["l"],
+  xl: reservoirSpacingValues["xl"],
+  xxl: reservoirSpacingValues["xxl"],
+  xxxl: reservoirSpacingValues["xxxl"],
+  // Reservoir semantic tokens
+  input: {
+    group: {
+      button: {
+        hstack: reservoirSpacingValues["xs"],
+        vstack: reservoirSpacingValues["xs"],
+      },
+      default: {
+        hstack: reservoirSpacingValues["l"],
+        vstack: reservoirSpacingValues["s"],
+      },
+    },
+  },
+  inset: {
+    default: reservoirSpacingValues["s"],
+    extranarrow: reservoirSpacingValues["xxs"],
+    extrawide: reservoirSpacingValues["xl"],
+    narrow: reservoirSpacingValues["xs"],
+    wide: reservoirSpacingValues["l"],
+  },
+  grid: {
+    default: reservoirSpacingValues["l"],
+    xxs: reservoirSpacingValues["xxs"],
+    xs: reservoirSpacingValues["xs"],
+    s: reservoirSpacingValues["s"],
+    m: reservoirSpacingValues["m"],
+    l: reservoirSpacingValues["l"],
+    xl: reservoirSpacingValues["xl"],
+    xxl: reservoirSpacingValues["xxl"],
+  },
+  page: {
+    hstack: reservoirSpacingValues["l"],
+    vstack: reservoirSpacingValues["l"],
+  },
+  table: {
+    column: reservoirSpacingValues["m"],
+  },
+};
+export const spacing = {
+  ...chakraSpacingTokens,
+  ...reservoirSpacingTokens,
 };